PP-77 | Front-End | Change the color of the requests and offers!

+++ src/sql/path_pals_create.sql
@@ -228,14 +228,14 @@ INSERT INTO persons(p_first_name, p_last_name, p_genre)
 VALUES ('Rúben', 'Costa', 'M');
 
 INSERT INTO users(u_username, u_email, p_id)
-VALUES('Cavalo47', 'cavalo47@gmail.com', LAST_INSERT_ID());
+VALUES('userTest47', 'userTest47@gmail.com', LAST_INSERT_ID());
 
 SET @u_id = (
 	SELECT u_id FROM users AS u WHERE u.u_id = LAST_INSERT_ID()
 );
 
 INSERT INTO passwords(pw_hashed_password, u_id)
-VALUES('ruben3947583763826', @u_id);
+VALUES('admin123admin', @u_id);
 
 SET @ut_admin_id = (
 	SELECT ut_id FROM user_types AS ut WHERE ut.ut_type = 'admin'
@@ -254,14 +254,14 @@ INSERT INTO persons(p_first_name, p_last_name, p_genre)
 VALUES ('Rodrigo', 'Costa', 'M');
 
 INSERT INTO users(u_username, u_email, p_id)
-VALUES('Cavalo45', 'cavalo45@gmail.com', LAST_INSERT_ID());
+VALUES('userTest45', 'userTest45@gmail.com', LAST_INSERT_ID());
 
 SET @u_id = (
 	SELECT u_id FROM users AS u WHERE u.u_id = LAST_INSERT_ID()
 );
 
 INSERT INTO passwords(pw_hashed_password, u_id)
-VALUES('rodrigo2344', @u_id);
+VALUES('123', @u_id);
 
 SET @ut_director_id = (
 	SELECT ut_id FROM user_types AS ut WHERE ut.ut_type = 'director'
@@ -280,14 +280,14 @@ INSERT INTO persons(p_first_name, p_last_name, p_genre)
 VALUES ('Daniel', 'Costa', 'M');
 
 INSERT INTO users(u_username, u_email, p_id)
-VALUES('Cavalo46', 'cavalo46@gmail.com', LAST_INSERT_ID());
+VALUES('userTest46', 'userTest46@gmail.com', LAST_INSERT_ID());
 
 SET @u_id = (
 	SELECT u_id FROM users AS u WHERE u.u_id = LAST_INSERT_ID()
 );
 
 INSERT INTO passwords(pw_hashed_password, u_id)
-VALUES('daniel45555', @u_id);
+VALUES('123', @u_id);
 
 SET @ut_user_id = (
 	SELECT ut_id FROM user_types AS ut WHERE ut.ut_type = 'user'
@@ -346,14 +346,14 @@ INSERT INTO persons(p_first_name, p_last_name, p_genre)
 VALUES ('João', 'Machado', 'M');
 
 INSERT INTO users(u_username, u_email, p_id)
-VALUES('Cavalo49', 'cavalo49@gmail.com', LAST_INSERT_ID());
+VALUES('userTest49', 'userTest49@gmail.com', LAST_INSERT_ID());
 
 SET @driver_u_id = (
 	SELECT u_id FROM users AS u WHERE u.u_id = LAST_INSERT_ID()
 );
 
 INSERT INTO passwords(pw_hashed_password, u_id)
-VALUES('joao492498', @u_id);
+VALUES('123', @u_id);
 
 SET @ut_user_id = (
 	SELECT ut_id FROM user_types AS ut WHERE ut.ut_type = 'user'
@@ -786,7 +786,11 @@ DELIMITER ;
 DELIMITER $$
 CREATE PROCEDURE get_all_offers()
 BEGIN
-    SELECT * FROM all_offers_view;
+    SELECT * FROM all_offers_view AS aov 
+    INNER JOIN tickets AS t ON t.t_id = aov.ticket_id
+    INNER JOIN ticket_status AS ts ON ts.ts_id = t.ts_id
+    INNER JOIN users AS u ON u.u_id = t.u_id
+    WHERE ts.ts_status != 'open' AND ts.ts_status != 'close';
 END $$
 DELIMITER ;
 
@@ -795,7 +799,12 @@ DELIMITER ;
 DELIMITER $$
 CREATE PROCEDURE get_all_requests()
 BEGIN
-    SELECT * FROM all_requested_view;
+    SELECT *
+    FROM all_requested_view AS arv
+    INNER JOIN tickets AS t ON t.t_id = arv.ticket_id
+    INNER JOIN ticket_status AS ts ON ts.ts_id = t.ts_id
+    INNER JOIN users AS u ON u.u_id = t.u_id
+    WHERE ts.ts_status != 'open' AND ts.ts_status != 'close';
 END $$
 DELIMITER ;
 
@@ -804,7 +813,11 @@ DELIMITER ;
 DELIMITER $$
 CREATE PROCEDURE get_user_offers(user_email VARCHAR(255))
 BEGIN
-    SELECT * FROM all_offers_view AS aov WHERE aov.user_email = user_email;
+    SELECT * FROM all_offers_view AS aov 
+    INNER JOIN tickets AS t ON t.t_id = aov.ticket_id
+    INNER JOIN ticket_status AS ts ON ts.ts_id = t.ts_id
+    INNER JOIN users AS u ON u.u_id = t.u_id
+    WHERE aov.user_email = user_email AND ts.ts_status != 'open' AND ts.ts_status != 'close';
 END $$
 DELIMITER ;
 
@@ -812,8 +825,13 @@ DELIMITER ;
 
 DELIMITER $$
 CREATE PROCEDURE get_user_requests(user_email VARCHAR(255))
-BEGIN
-    SELECT * FROM all_requested_view AS arv WHERE arv.user_email = user_email;
+BEGIN    
+    SELECT *
+    FROM all_requested_view AS arv
+    INNER JOIN tickets AS t ON t.t_id = arv.ticket_id
+    INNER JOIN ticket_status AS ts ON ts.ts_id = t.ts_id
+    INNER JOIN users AS u ON u.u_id = t.u_id
+    WHERE u.u_email = user_email AND ts.ts_status != 'open' AND ts.ts_status != 'close';
 END $$
 DELIMITER ;
 
@@ -869,54 +887,58 @@ DELIMITER $$
 CREATE PROCEDURE accept_offer(r_id INT, accepting_user_email VARCHAR(255))
 BEGIN
      SET @traveler_id = (
-        SELECT u_id 
-        FROM users 
-        WHERE u_email = accepting_user_email
+        SELECT u_id FROM users WHERE u_email = accepting_user_email
     );
 
     SET @num_accepted = (
-        SELECT COUNT(*)
-        FROM connections
-        WHERE u_id_traveler = @traveler_id
+        SELECT COUNT(*) FROM connections WHERE u_id_driver = @driver_id
     );
 
     IF @num_accepted > 0 THEN
-        SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'O usuário já aceitou uma oferta';
+        SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'O utilizador já aceitou uma oferta';
     ELSE
-        SET @driver_id = (
+        /*SET @driver_id = (
             SELECT t.u_id FROM rides AS r
             INNER JOIN tickets AS t ON t.t_id = r.t_id
             INNER JOIN ride_types AS rt ON rt.rt_id = r.rt_id
             WHERE r.r_id = r_id AND rt.rt_type = 'proposed'
+        );*/
+        
+        # ------
+        SET @driver_id = (
+            SELECT t.u_id
+            FROM rides AS r
+            INNER JOIN tickets AS t ON t.t_id = r.t_id
+            WHERE r.r_id = r_id
         );
-
+        # ------
+        
         SET @ticket_id = (
-			SELECT t.t_id 
+            SELECT t.t_id 
             FROM tickets AS t
-            WHERE t.u_id = @driver_id
+            INNER JOIN ticket_status AS ts ON ts.ts_id = t.ts_id
+            INNER JOIN rides AS r ON r.t_id = t.t_id
+            WHERE t.u_id = @driver_id AND r.r_id = r_id AND ts.ts_status != 'open' AND ts.ts_status != 'close'
         );
 
         INSERT INTO connections(r_id, u_id_driver, u_id_traveler)
         VALUES(r_id, @driver_id, @traveler_id);
         
         SET @connections_count = (
-			SELECT COUNT(*)
+            SELECT COUNT(*)
             FROM connections AS c
             WHERE c.c_id = LAST_INSERT_ID()
         );
         
         SET @ticket_status_open_id = (
-			SELECT ts.ts_id FROM tickets AS t 
-            INNER JOIN ticket_status AS ts ON ts.ts_id = t.ts_id
+            SELECT ts.ts_id 
+            FROM ticket_status AS ts
             WHERE ts.ts_status = 'open'
         );
         
-        IF @connections_count > 0 THEN
-			UPDATE tickets AS t
-            SET
-				t.ts_id = @ticket_status_open_id
-			WHERE t.t_id = @ticket_id;
-		END IF;
+		UPDATE tickets AS t
+		SET t.ts_id = @ticket_status_open_id
+		WHERE t.t_id = @ticket_id;
     END IF;
 END $$
 DELIMITER ;
@@ -927,9 +949,9 @@ DELIMITER $$
 CREATE PROCEDURE accept_request(r_id INT, accepting_user_email VARCHAR(255))
 BEGIN
     SET @driver_id = (
-            SELECT u_id FROM users WHERE u_email = accepting_user_email
-        );
-
+        SELECT u_id FROM users WHERE u_email = accepting_user_email
+    );
+    
     SET @num_accepted = (
         SELECT COUNT(*) FROM connections WHERE u_id_driver = @driver_id
     );
@@ -938,38 +960,39 @@ BEGIN
         SIGNAL SQLSTATE '45000' SET MESSAGE_TEXT = 'O utilizador já aceitou uma oferta ou um pedido';
     ELSE
         SET @traveler_id = (
-			SELECT t.u_id FROM rides AS r
+            SELECT t.u_id
+            FROM rides AS r
             INNER JOIN tickets AS t ON t.t_id = r.t_id
-            INNER JOIN ride_types AS rt ON rt.rt_id = r.rt_id
-            WHERE r.r_id = r_id AND rt.rt_type = 'requested'
+            WHERE r.r_id = r_id
         );
         
         SET @ticket_id = (
-			SELECT t.t_id 
+            SELECT t.t_id 
             FROM tickets AS t
-            WHERE t.u_id = @traveler_id
+            INNER JOIN ticket_status AS ts ON ts.ts_id = t.ts_id
+            INNER JOIN rides AS r ON r.t_id = t.t_id
+            WHERE t.u_id = @traveler_id AND r.r_id = r_id AND ts.ts_status != 'open' AND ts.ts_status != 'close'
         );
-
+        
         INSERT INTO connections(r_id, u_id_traveler, u_id_driver)
         VALUES(r_id, @traveler_id, @driver_id);
         
         SET @connections_count = (
-			SELECT COUNT(*)
+            SELECT COUNT(*)
             FROM connections AS c
             WHERE c.c_id = LAST_INSERT_ID()
         );
         
         SET @ticket_status_open_id = (
-			SELECT ts.ts_id FROM tickets AS t 
-            INNER JOIN ticket_status AS ts ON ts.ts_id = t.ts_id
+            SELECT ts.ts_id 
+            FROM ticket_status AS ts
             WHERE ts.ts_status = 'open'
         );
         
         IF @connections_count > 0 THEN
-			UPDATE tickets AS t
-            SET
-				t.ts_id = @ticket_status_open_id
-			WHERE t.t_id = @ticket_id;
+            UPDATE tickets AS t
+            SET t.ts_id = @ticket_status_open_id
+            WHERE t.t_id = @ticket_id;
 		END IF;
     END IF;
 END $$
@@ -990,10 +1013,42 @@ DELIMITER $$
 CREATE PROCEDURE cancel_requests(user_email VARCHAR(255))
 BEGIN
     SET @u_id = (
-		SELECT u_id FROM users WHERE u_email = user_email
-        );
+        SELECT u_id FROM users WHERE u_email = user_email
+    );
+    
+    SET @connections_traveler_count = (
+        SELECT COUNT(*) FROM connections WHERE u_id_traveler = @u_id
+    );
+    
+    SET @connections_driver_count = (
+        SELECT COUNT(*) FROM connections WHERE u_id_driver = @u_id
+    );
+    
+    SET @ticket_id = (
+        SELECT t.t_id
+        FROM connections AS c
+        INNER JOIN rides AS r ON r.r_id = c.r_id
+        INNER JOIN tickets AS t ON t.t_id = r.t_id
+        WHERE u_id_driver = @u_id OR u_id_traveler = @u_id
+    );
 
-    DELETE FROM connections WHERE u_id_traveler = @u_id;
+    IF @connections_traveler_count > 0 THEN
+        DELETE FROM connections WHERE u_id_traveler = @u_id;
+    END IF;
+    
+    IF @connections_driver_count > 0 THEN
+        DELETE FROM connections WHERE u_id_driver = @u_id;
+    END IF;
+    
+    SET @ticket_status_close_id = (
+        SELECT ts.ts_id 
+        FROM ticket_status AS ts
+        WHERE ts.ts_status = 'close'
+    );
+    
+    UPDATE tickets AS t
+    SET t.ts_id = @ticket_status_close_id
+    WHERE t.t_id = @ticket_id;
 END $$
 DELIMITER ;
 
@@ -1003,10 +1058,42 @@ DELIMITER $$
 CREATE PROCEDURE cancel_offers(user_email VARCHAR(255))
 BEGIN
     SET @u_id = (
-		SELECT u_id FROM users WHERE u_email = user_email
-        );
+        SELECT u_id FROM users WHERE u_email = user_email
+    );
+    
+    SET @connections_traveler_count = (
+        SELECT COUNT(*) FROM connections WHERE u_id_traveler = @u_id
+    );
+    
+    SET @connections_driver_count = (
+        SELECT COUNT(*) FROM connections WHERE u_id_driver = @u_id
+    );
+    
+    SET @ticket_id = (
+        SELECT t.t_id
+        FROM connections AS c
+        INNER JOIN rides AS r ON r.r_id = c.r_id
+        INNER JOIN tickets AS t ON t.t_id = r.t_id
+        WHERE u_id_driver = @u_id OR u_id_traveler = @u_id
+    );
 
-    DELETE FROM connections WHERE u_id_driver = @u_id;
+    IF @connections_traveler_count > 0 THEN
+        DELETE FROM connections WHERE u_id_traveler = @u_id;
+    END IF;
+    
+    IF @connections_driver_count > 0 THEN
+        DELETE FROM connections WHERE u_id_driver = @u_id;
+    END IF;
+    
+    SET @ticket_status_close_id = (
+        SELECT ts.ts_id 
+        FROM ticket_status AS ts
+        WHERE ts.ts_status = 'close'
+    );
+    
+    UPDATE tickets AS t
+    SET t.ts_id = @ticket_status_close_id
+    WHERE t.t_id = @ticket_id;
 END $$
 DELIMITER ;
 
@@ -1030,7 +1117,9 @@ BEGIN
 
     IF EXISTS (SELECT * FROM rides WHERE r_id = ride_id AND t_id IN (SELECT t_id FROM tickets WHERE u_id = @u_id)) THEN
 		SET @ticket_id = (
-			SELECT t.t_id FROM tickets AS t WHERE t.u_id = @u_id
+			SELECT t.t_id FROM tickets AS t 
+            INNER JOIN rides AS r ON r.t_id = t.t_id
+            WHERE t.u_id = @u_id AND r.r_id = ride_id
         );
         
         SET @ticket_status_close_id = (
@@ -1063,7 +1152,9 @@ BEGIN
 
     IF EXISTS (SELECT * FROM rides WHERE r_id = ride_id AND t_id IN (SELECT t_id FROM tickets WHERE u_id = @u_id)) THEN
 		SET @ticket_id = (
-			SELECT t.t_id FROM tickets AS t WHERE t.u_id = @u_id
+			SELECT t.t_id FROM tickets AS t 
+            INNER JOIN rides AS r ON r.t_id = t.t_id
+            WHERE t.u_id = @u_id AND r.r_id = ride_id
         );
         
         SET @ticket_status_close_id = (
